Add `[` and `]` to OSK layer
For easier access on super smol boards that use R2_FOUR (pain27, agony, kawaii). Add `WF` combo to more easily access that layer.

Also align `-` and `=` symbols on R2_FOUR to align with other layouts. Adapting to this change on these small boards was disorientating and I was making a lot of small silly mistakes.